Divide archivo ui en dos archivos para simplificar lectura del código
